Move dashboard code to controller
